requirements.txt
dotenv
streamlit 
google-generativeai
psycopg2-binary
bcrypt
pyjwt